--- backend/spring-boot-3/src/main/java/back/springbootdeveloper/seungchan/exception/user/UserNotExistException.java
@@ -1,4 +1,0 @@
-package back.springbootdeveloper.seungchan.exception.user;
-
-public class UserNotExistException extends RuntimeException{
-}
