+++ backend/spring-boot-3/src/main/java/back/springbootdeveloper/seungchan/filter/exception/user/UserNotExistException.java
@@ -1,0 +1,4 @@
+package back.springbootdeveloper.seungchan.filter.exception.user;
+
+public class UserNotExistException extends RuntimeException {
+}
