+++ backend/spring-boot-3/src/main/java/back/springbootdeveloper/seungchan/service/LoginService.java
@@ -4,8 +4,7 @@ import back.springbootdeveloper.seungchan.entity.UserInfo;
 import back.springbootdeveloper.seungchan.entity.UserUtill;
 import back.springbootdeveloper.seungchan.dto.request.UserLoginRequest;
 import back.springbootdeveloper.seungchan.dto.response.UserLoginResponse;
-import back.springbootdeveloper.seungchan.exception.common.EmptyValueExistException;
-import back.springbootdeveloper.seungchan.exception.user.UserNotExistException;
+import back.springbootdeveloper.seungchan.filter.exception.user.UserNotExistException;
 import back.springbootdeveloper.seungchan.repository.UserRepository;
 import back.springbootdeveloper.seungchan.repository.UserUtilRepository;
 import jakarta.servlet.http.HttpServletRequest;
